ITS ALIVE
I've made the car along with the code that is needed for it to move, drift, and boost as one of some abilities that i want. IT EVEN HAS COLLISION. does it work? kinda, it moves but not how i intended but no worries. this is after resetting ~4 hours of progress on purpose as nothing that i made worked and it felt scammy to try keeping it. like gideon the all knowing ive read so many different programs and documents and hob cobbled this together