Fixed the Engine
The issue was that the sound file itself was bad. I thought it looped perfectly but it was off slightly. No code can fix that so I had to just find a new audio. You can find my wasted code in the other branch which is where any code that wasn't needed will go. Changed the boost