Audio Menu (WIP3)
Finally got the menu done. Tutorial I was using was using hboxcontainers from I swapped to a node for the organization and it finally let me adjust the label and slider manually. Adding more to the map instead of it being a loop. Also fixed a bug where when a shiba sound plays there was a chance the menu just didn't load. Also no "Back Button" for a bit, maybe forever, as it always applies no matter what. Just need the music slider to control the music and the sfx slider the sfx.