Stuff in General
Still working on the Audio Options. Got the button to play a return sound. Currently working on making better map but more importantly working on the steering as someone mentioned they disliked it which is fair cause so did I. The solution to it is weird so check the code for that. I'ma finish the map and after that add the menu to the game and music to it as well.