Second Demo
Added a sound in the world which caused all audio to increase greatly. Had to lower the car engine sound with an editor. Added a far better map than before. I adjusted the steering as it felt bad but I need feedback for that to see if people like it.