UI (Done)
Do not know why but after reverting and trying a slightly different way the speed and timer stopped causing null errors. The Boost bar, speed, and timer are shown at the top left. I have the code to add trackers to laps and such, even adding it a .txt file but that will come next commit, maybe.