UI (Extra Done)
I forgot that I in fact was not done. I have some buttons that allow be to test the timer but I have made a scene that when the player collides it trackers a lap which is great. I am going to be making a character select next. I think.