Y'know, Stuff
Added a working checkpoint system, fixed the boost working when empty,  and made it so speed is gradually lost not instantly cut when boosting is stopped. I am currently creating a car select for future cars but want to finished the script before the demo. Also sleepy so ima work later.