End Menu (WIP3)
While it functions, I am attemtping to make the play again and quit button create sounds. The main issue I am trying to fix is the sounds doesn't play unless the scene is unfrozen but that looks off so I have been trying for a work around. Should be done next commit.